Created a new file for logistic regression and created a file for training and testing of data
